LayoutTests/security/cannot-read-self-from-file.html

<!DOCTYPE html>
<html>
<head>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
testRunner.setAllowUniversalAccessFromFileURLs(false);
testRunner.setAllowFileAccessFromFileURLs(false);

var t = async_test('file: should be a unique-origin protocol for XHR purposes');
window.addEventListener('message', t.step_func(function(evt) {
    assert_equals(evt.data, 'SUCCESS');
    t.done();
}));
</script>
</head>
<body>
<iframe src="resources/cannot-read-self-from-file.html"></iframe>
Documents loaded from file: shouldn't be able to access themselves via XHR.
</body>
</html>
